Allow for easy version tracking
This definitely isn't the best way to do it, but it's a start. Each time we make a commit, we need to update kCOMMIT and ensure that kBRANCH is correct. Because of the way that we are doing this, kCOMMIT in GitHub will always be one behind the actual commit, but we just need to remember to change the value once we pull.